+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Labels: 
     - fix wide text
     - remove shape hovering
+- RubberBand: fix rubberband deletion when putting it outside view point
+
 
 ## [0.25.0]
 ### Add