+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.26.0]
+<<<<<<< HEAD
 ### Add
 - RemoteFigure.setFeatureFilter to directly edit rubberbands' value from external requests
 - Multiplot.setFeatureFilter to directly edit rubberbands' value from external requests
@@ -14,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Labels: 
     - fix wide text
     - remove shape hovering
+=======
+### Fix
+- RubberBand: fix rubberband deletion when putting it outside view point
+
+>>>>>>> origin/dev
 
 ## [0.25.0]
 ### Add