+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 - RemoteFigure.setFeatureFilter to directly edit rubberbands' value from external requests
 - Multiplot.setFeatureFilter to directly edit rubberbands' value from external requests
+- Emit rubberband changes
 
 ### Fix
 - Labels: 
     - fix wide text
     - remove shape hovering
+- RubberBand: fix rubberband deletion when putting it outside view point
 
 ## [0.25.0]
 ### Add