+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 - RemoteFigure.setFeatureFilter to directly edit rubberbands' value from external requests
 - Multiplot.setFeatureFilter to directly edit rubberbands' value from external requests
+- Emit rubberband changes
 
 ### Fix
 - Labels: