+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - fix wide text
     - remove shape hovering
 - RubberBand: fix rubberband deletion when putting it outside view point
+- Fixing tests on dev
 
 ## [0.25.1]
 ### Doc